Button Updates
Made the buttons stay pressed until moused off of or m1 is released and generally has more graceful state transitions. Separated button rendering from state changes and moved the rendering post state updates/sounds so that it's easier to switch to custom buttons when ready and so that rendering per state is consistent.