Multiplayer Fixes
Fixed an issue with using `Sync` in an `onInit` referencing `onSetPlayer` when setting character names which would break the client. The client now just requests a name if the character doesnt have a first name. Fixed an issue with button thrashing. The button state wouldnt update and the buttons were broken when attached to `CPlayer` so i attached it to `CRules` instead. Started work on custom scoreboard