Bug Fix
There was an issue that took wayyyyyy too long to fix. The issue was the player list displays would not move back up the list if a char dies. This fixes that in a roundabout way (the current version has the server send a command to the client that you need to scroll a specific list up)