Scoreboard Tweaks, Display Mode Toggle
Added a toggle for a smaller display mode for the multicharacter UI, cycling from full frames to simple frames to no frame then back to full frame. Changed the style of the scoreboard to fit in with the theme better. Forced the menu font to a specific style to prevent flickering when opening up the scoreboard.